requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-docx==0.8.11
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
urllib3==1.26.5
Werkzeug==2.0.1
xlrd==2.0.1
